Add support for choices

+++ crates/pod/src/en/encode.rs
@@ -26,7 +26,7 @@ pub trait Encode: Sized + self::sealed::Sealed {
     const TYPE: Type;
 
     /// The size in bytes of the encoded value.
-    fn size(&self) -> usize;
+    fn size(&self) -> u32;
 
     /// Encode the value into the writer.
     fn encode(&self, writer: impl Writer) -> Result<(), Error>;
@@ -52,7 +52,7 @@ impl Encode for bool {
     const TYPE: Type = Type::BOOL;
 
     #[inline]
-    fn size(&self) -> usize {
+    fn size(&self) -> u32 {
         4
     }
 
@@ -92,7 +92,7 @@ where
     const TYPE: Type = Type::ID;
 
     #[inline]
-    fn size(&self) -> usize {
+    fn size(&self) -> u32 {
         4
     }
 
@@ -124,7 +124,7 @@ impl Encode for i32 {
     const TYPE: Type = Type::INT;
 
     #[inline]
-    fn size(&self) -> usize {
+    fn size(&self) -> u32 {
         4
     }
 
@@ -156,7 +156,7 @@ impl Encode for i64 {
     const TYPE: Type = Type::LONG;
 
     #[inline]
-    fn size(&self) -> usize {
+    fn size(&self) -> u32 {
         8
     }
 
@@ -190,7 +190,7 @@ impl Encode for f32 {
     const TYPE: Type = Type::FLOAT;
 
     #[inline]
-    fn size(&self) -> usize {
+    fn size(&self) -> u32 {
         4
     }
 
@@ -222,7 +222,7 @@ impl Encode for f64 {
     const TYPE: Type = Type::DOUBLE;
 
     #[inline]
-    fn size(&self) -> usize {
+    fn size(&self) -> u32 {
         8
     }
 
@@ -256,7 +256,7 @@ impl Encode for Rectangle {
     const TYPE: Type = Type::RECTANGLE;
 
     #[inline]
-    fn size(&self) -> usize {
+    fn size(&self) -> u32 {
         8
     }
 
@@ -288,7 +288,7 @@ impl Encode for Fraction {
     const TYPE: Type = Type::FRACTION;
 
     #[inline]
-    fn size(&self) -> usize {
+    fn size(&self) -> u32 {
         8
     }
 
@@ -321,8 +321,8 @@ impl<const N: usize> Encode for [u8; N] {
     const TYPE: Type = Type::BYTES;
 
     #[inline]
-    fn size(&self) -> usize {
-        N
+    fn size(&self) -> u32 {
+        N as u32
     }
 
     #[inline]
@@ -355,7 +355,7 @@ impl Encode for Pointer {
     const TYPE: Type = Type::POINTER;
 
     #[inline]
-    fn size(&self) -> usize {
+    fn size(&self) -> u32 {
         16
     }
 
@@ -397,7 +397,7 @@ impl Encode for Fd {
     const TYPE: Type = Type::FD;
 
     #[inline]
-    fn size(&self) -> usize {
+    fn size(&self) -> u32 {
         8
     }
 
@@ -438,7 +438,7 @@ where
     const TYPE: Type = T::TYPE;
 
     #[inline]
-    fn size(&self) -> usize {
+    fn size(&self) -> u32 {
         EncodeUnsized::size(*self)
     }
